requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.7
pandas>=1.3.0
numpy>=1.21.0
pytest>=7.4.3
pytest-cov>=2.12.0
responses>=0.13.0
black>=23.11.0
flake8>=6.1.0
tqdm>=4.62.0
python-dateutil>=2.8.2
dataclasses>=0.6
typing-extensions>=3.7.4 